Introduce 'funcall' and a special form 'function'
funcall is a function that knows how to apply functions.

Fallout from funcall is that we need a way to return the functional
value of a function, so as to prevent an unintential evaluation of the
function.

The function funcall lets us take the result of a calculation coming
from the parametric world and put it into the function world as a value.