Add section 2.4 Comparing lisp(1) and lisp(2)
Inroduce the function f.lookup to search for a name in the function
environment. If the name cannot be found, a function calling wrong will
be returned.

This device also introduces a kind of deferred error since such an error
does not occur in the reference to the non-existing function, but rather
it occurs in the application of the function, which could occur later or
perhaps not at all.